# performance opt: using copy instead of append